Added list function: each()
