+++ lib/less/functions/index.js
@@ -9,6 +9,7 @@ module.exports = function(environment) {
     require("./color");
     require("./color-blending");
     require("./data-uri")(environment);
+    require("./list");
     require("./math");
     require("./number");
     require("./string");
